Move movie detail view model to be individual file
