+++ app/src/main/java/com/enginebai/moviehunt/ui/detail/MovieDetailFragment.kt
@@ -4,32 +4,19 @@ import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
 import android.util.Log
-import android.view.MenuItem
 import android.view.View
-import androidx.appcompat.app.AppCompatActivity
 import androidx.core.os.bundleOf
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.Transformations
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.enginebai.base.view.BaseFragment
-import com.enginebai.base.view.BaseViewModel
 import com.enginebai.moviehunt.R
-import com.enginebai.moviehunt.data.local.MovieModel
 import com.enginebai.moviehunt.data.local.getPosterUrlWithLargeSize
-import com.enginebai.moviehunt.data.remote.CastListing
-import com.enginebai.moviehunt.data.remote.Review
-import com.enginebai.moviehunt.data.remote.Video
-import com.enginebai.moviehunt.data.repo.MovieRepo
 import com.enginebai.moviehunt.ui.MovieClickListener
 import com.enginebai.moviehunt.utils.loadImage
 import com.enginebai.moviehunt.utils.openFragment
-import io.reactivex.schedulers.Schedulers
 import kotlinx.android.synthetic.main.fragment_movie_detail.*
 import kotlinx.android.synthetic.main.view_toolbar.*
 import org.koin.androidx.viewmodel.ext.android.viewModel
-import org.koin.core.component.inject
 
 class MovieDetailFragment : BaseFragment(), MovieClickListener {
 
@@ -99,86 +86,3 @@ class MovieDetailFragment : BaseFragment(), MovieClickListener {
     }
 }
 
-class MovieDetailViewModel : BaseViewModel() {
-    private val movieRepo: MovieRepo by inject()
-
-    private val _movieDetail = MutableLiveData<MovieModel>()
-    val movieDetail: LiveData<MovieModel> = _movieDetail
-    private val _movieVideos = MutableLiveData<List<Video>>()
-    val videos: LiveData<List<Video>> = _movieVideos
-    private val _movieReviews = MutableLiveData<List<Review>>()
-    val review: LiveData<Review?> = Transformations.map(_movieReviews) {
-        it.findBestReview()
-    }
-    private val _movieCasts = MutableLiveData<List<CastListing.Cast>>()
-    val casts: LiveData<List<CastListing.Cast>> = _movieCasts
-    private val _similarMovies = MutableLiveData<List<MovieModel>>()
-    val similarMovies: LiveData<List<MovieModel>> = _similarMovies
-    private val _recommendationMovies = MutableLiveData<List<MovieModel>>()
-    val recommendationMovies: LiveData<List<MovieModel>> = _recommendationMovies
-
-    fun fetchMovieDetail(id: String) {
-        movieRepo.fetchMovieDetail(id)
-            .subscribeOn(Schedulers.io())
-            .subscribe()
-            .disposeOnCleared()
-
-        movieRepo.fetchMovieVideos(id)
-            .subscribeOn(Schedulers.io())
-            .doOnSuccess {
-                _movieVideos.postValue(it)
-            }
-            .subscribe()
-            .disposeOnCleared()
-
-        movieRepo.fetchMovieReviews(id)
-            .subscribeOn(Schedulers.io())
-            .doOnSuccess {
-                _movieReviews.postValue(it)
-            }
-            .subscribe()
-            .disposeOnCleared()
-
-        movieRepo.fetchMovieCasts(id)
-            .subscribeOn(Schedulers.io())
-            .doOnSuccess {
-                _movieCasts.postValue(it)
-            }
-            .subscribe()
-            .disposeOnCleared()
-
-        movieRepo.fetchSimilarMovies(id)
-            .subscribeOn(Schedulers.io())
-            .doOnSuccess {
-                _similarMovies.postValue(it)
-            }
-            .subscribe()
-            .disposeOnCleared()
-
-        movieRepo.fetchRecommendationMovies(id)
-            .subscribeOn(Schedulers.io())
-            .doOnSuccess {
-                _recommendationMovies.postValue(it)
-            }
-            .subscribe()
-            .disposeOnCleared()
-
-        movieRepo.getMovieDetail(id)
-            .doOnNext {
-                _movieDetail.postValue(it)
-            }.subscribe()
-            .disposeOnCleared()
-    }
-
-    private fun List<Review>.findBestReview(): Review? {
-        return filter {
-            !it.content.isNullOrBlank() &&
-                    !it.author?.avatarPath.isNullOrBlank() &&
-                    it.author?.rating != null
-        }.maxByOrNull { it.author!!.rating!! }
-            ?: this.firstOrNull {
-                !it.content.isNullOrBlank() &&
-                        !it.author?.avatarPath.isNullOrBlank()
-            }
-    }
-}
